creacion del paquete product
se creo el paquete product para hacer una APIRest

+++ sushi/src/main/java/com/konoha/sushi/user/application/UserServiceImpl.java
@@ -46,6 +46,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void delete(Long id) {
+        if(!userRepository.existsById(id)){
+            throw new ResourceNotFoundException("User not found with id: " + id);
+        }
         userRepository.deleteById(id);
     }
 }
